Extract utility functions to utility modules
