+++ __init__.py
@@ -1,0 +1,2 @@
+__name__ = 'synergile'
+__package__ = 'synergile'
