+++ bot.py
@@ -3,10 +3,10 @@ import asyncio
 import datetime
 from datetime import datetime
 from datetime import timezone
-import re
 import youtube_dl
 from discord.ext import commands
-from discord.ext import tasks
+from util.pyutil import buildMultiMatchString, splitArgs
+from util.discordutil import resolveMember, modActionLogEmbed
 import os
 import math
 import random
@@ -14,14 +14,13 @@ desc= "Moderation bot engineered by CodeWritten, wakfi, and jedi3"
 bot = commands.Bot(command_prefix='$', case_insensitive=True, description=desc)
 bot.remove_command('help') #removing the default help cmd
 #NO_MENTIONS = discord.AllowedMentions(everyone=False,users=False,roles=False) - add in d.py 1.4
-SNOWFLAKE_REGEX = re.compile('\D') #compile regular expression matching all characters that aren't digits
 
 
 @bot.event
 async def on_ready():
     await bot.change_presence(status=discord.Status.online, activity=discord.Game(f'{bot.command_prefix}help for commands'))
     print (f"Bot online")
-
+    
 @bot.event
 async def on_command_error(ctx, error):
     await ctx.send("An error occured!\n```{}```".format(error))
@@ -43,7 +42,7 @@ async def profile(ctx, *, member= None):
     
     #generate profile embed and send
     if(isinstance(mem, list)):
-        usersFound = buildMultiMatchString('profile', mem, member)
+        usersFound = buildMultiMatchString(bot.command_prefix, 'profile', mem, member)
         await ctx.send(usersFound)
     else:
         embed = profileEmbed(ctx.message.author, mem)
@@ -101,7 +100,7 @@ async def kick(ctx,*, member=None, reason="No reason provided"):
         return
     
     if(isinstance(mem, list)):
-        usersFound = buildMultiMatchString('kick', mem, member)
+        usersFound = buildMultiMatchString(bot.command_prefix, 'kick', mem, member)
         await ctx.send(usersFound)
     else:
         indexReason = -1
@@ -141,7 +140,7 @@ async def ban(ctx,*, member=None, reason = "No reason provided"):
         return
     
     if(isinstance(mem, list)):
-        usersFound = buildMultiMatchString('ban', mem, member)
+        usersFound = buildMultiMatchString(bot.command_prefix, 'ban', mem, member)
         await ctx.send(usersFound)
     else:
         indexReason = -1
@@ -237,79 +236,6 @@ async def help(ctx):
 @bot.command(desc="Says pong!")
 async def ping(ctx):
     await ctx.send('Pong! {} ms'.format(bot.latency*1000))
-
-#should be a util when cog setup
-#resolve a string to a member object
-async def resolveMember(ctx, stringToResolve):
-    mc = commands.MemberConverter()
-    stringToResolve = stringToResolve.strip()
-    if isSnowflake(stringToResolve): 
-        #gave an number that may be an ID
-        member = ctx.guild.get_member(int(stringToResolve)) #if stringToResolve is not a valid snowflake, mem=None
-    elif '@' in stringToResolve:
-        #mentioned a user
-        try:
-            member = await mc.convert(ctx, stringToResolve)
-        except Exception:
-            member = None
-    else:
-        try:
-            memList = await ctx.guild.query_members(query = stringToResolve)
-        except Exception:
-            memList = []
-        if len(memList)==0:
-            member = None
-        elif len(memList)==1:
-            member = memList[0]
-        else:
-            return memList
-    return member
-
-#should be a util when cog setup
-def buildMultiMatchString(command,mem,member):
-    strBuilder=f'Found {len(mem)} possible matches for "{member}":```'
-    strBuilder=''.join([strBuilder, ''.join(f'\n{index+1}. {memMatch}' for index,memMatch in enumerate(mem))])
-    strBuilder+=f'```'
-    if len(mem)==5:
-        strBuilder+=f'\n(number of matches shown is capped at 5, there may or may not be more)'
-    strBuilder+=f'\nTry using the {bot.command_prefix}{command} command again with a more specific search term!'
-    return strBuilder
-    
-#should be a util class when cog setup
-def modActionLogEmbed(action,member,reason,issuer=bot):
-    embed = discord.Embed(colour = discord.Colour.red(), description = f'{action} {member.mention} with reason: {reason}', timestamp = datetime.now(timezone.utc))
-    embed.set_footer(text= f"Issued by {issuer}({issuer.id})", icon_url=issuer.avatar_url)
-    return embed
-    
-#helper
-#returns true if a value only contains characters 0-9, false otherwise. does not check length. 
-#a consequence of this is that is will return true on empty values
-def isSnowflake(snowflake):
-    return isinstance(snowflake,str) and not SNOWFLAKE_REGEX.search(snowflake)
-
-#do i even need to say that this is a util? this is a util
-def splitArgs(input):
-    optionRegex = re.compile(' -[a-zA-Z]+')
-    if(not re.search(optionRegex, input)):
-        return [input]
-    spaceSplit = re.split(r' ', input)
-    optionSplitArgs = optionRegex.split(input)
-    optionSplitArgs = [arg.strip() for arg in optionSplitArgs]
-    flagsRegex = re.compile('-')
-    rawFlags = list(filter(flagsRegex.match, spaceSplit))
-    joinedFlags = ''.join(rawFlags)
-    rawFlags = re.split(flagsRegex, joinedFlags)
-    joinedFlags = ''.join(rawFlags)
-    flags = re.split(r'', joinedFlags)
-    flags.pop()  # remove tailing empty string
-    flags.pop(0) # remove leading empty string
-    flags = [flag.lower() for flag in flags]
-    return [optionSplitArgs, flags]
-
-#replace occurances of a substring from the back
-def rreplace(s, old, new, occurrence):
-    li = s.rsplit(old, occurrence)
-    return new.join(li)
     
 @bot.command(desc="Displays build info")
 async def build_info(ctx, file_override=None):
