+++ commands/uptime.py
@@ -1,17 +1,20 @@
 import discord, datetime, time
 from discord.ext import commands
 
+#this lets you base your import from the Synergile folder to get to the util directory
+import sys
+from os import path
+sys.path.append( path.dirname( path.dirname( path.dirname(path.abspath(__file__)) ) ) )
+
+from Synergile.util.pyutil import rreplace
+
+
 start_time = time.time()
 start_time = datetime.datetime.utcnow()
 
 class Uptime(commands.Cog, name='Uptime'):
 	def __init__(self, bot):
 		self.bot = bot
-	
-    #replace occurances of a substring from the back
-	def rreplace(self, s, old, new, occurrence):
-		li = s.rsplit(old, occurrence)
-		return new.join(li)
     
 	@commands.command(desc='uptime')
 	async def uptime(self, ctx : commands.Context):
@@ -29,11 +32,11 @@ class Uptime(commands.Cog, name='Uptime'):
 			uptime_stamp += f'**{minutes}** {"minute" if minutes==1 else "minutes"}, '
 		if seconds:
 			uptime_stamp += f'**{seconds}** {"second" if seconds==1 else "seconds"}, '
-		uptime_stamp=self.rreplace(s=uptime_stamp, old=', ', new='', occurrence=1)
+		uptime_stamp=rreplace(s=uptime_stamp, old=', ', new='', occurrence=1)
 		if uptime_stamp.count(',') > 1:
-			uptime_stamp=self.rreplace(s=uptime_stamp, old=', ', new=', and ', occurrence=1)
+			uptime_stamp=rreplace(s=uptime_stamp, old=', ', new=', and ', occurrence=1)
 		elif uptime_stamp.count(',')==1:
-			uptime_stamp=self.rreplace(s=uptime_stamp, old=', ', new=' and ', occurrence=1)
+			uptime_stamp=rreplace(s=uptime_stamp, old=', ', new=' and ', occurrence=1)
 		embed = discord.Embed(title='Uptime', description=f'The bot, {self.bot.user.name}, has been online for {uptime_stamp}')
 		await ctx.send(embed=embed)
